BjoernTranslate/requirements.txt
Flask
python-docx
transliterate
flask-cors
pywin32
py_eureka_client
python-dotenv